updated corrections with Taha

+++ common-core/libft/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: alara-gu <alara-gu@student.42barcelon      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/26 16:23:09 by alara-gu          #+#    #+#             */
-/*   Updated: 2024/06/28 14:44:20 by alara-gu         ###   ########.fr       */
+/*   Updated: 2024/06/28 16:58:44 by alara-gu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
