+++ common-core/libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: alara-gu <alara-gu@student.42barcelon      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/26 16:08:33 by alara-gu          #+#    #+#             */
-/*   Updated: 2024/06/28 14:45:24 by alara-gu         ###   ########.fr       */
+/*   Updated: 2024/06/28 16:59:18 by alara-gu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
